+++ addOns/reports/CHANGELOG.md
@@ -4,8 +4,12 @@ All notable changes to this add-on will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
+### Changed
+- Default to same dir as plan if none specified.
+
 ### Fixed
 - Ensure the Sections' options are fully shown always in Generate Report Dialog (Issue 8259).
+- Replace env vars in URL when used for report file name.
 
 ## [0.27.0] - 2023-12-19
 ### Changed